Implement the 'pop' method.
This method works because the default ASC ordering of the TimeUUID clustering key puts the messages in a FIFO order when taken with a SELECT.
The returned entries are market (tombstoned) for the deletion.

Right now the "real" deletion, the repair, runs once a week... probably this tables must be combed more frequently.

Cc: @gbagnoli